+++ History.md
@@ -2,7 +2,7 @@
 
 ## 0.3.0
 
-* renamed from ChefDK::Template::Plugin to ChefGen::Template at Chef's request so as to not pollute the ChefDK namespace
+* renamed from ChefDK::Template::Plugin to ChefGen::Flavors at Chef's request so as to not pollute the ChefDK namespace
   * 0.2.0 exists only with the old name as a transitional to remind people that the name has changed
 
 ## 0.1.0